chatbot-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
langchain>=0.2.8
langchain-openai>=0.1.17
pydantic>=2.8.2
